pyproject.toml
[tool.poetry]
name = "pythondev"
version = "0.1.0"
description = ""
authors = ["Saeed Gholami <saeed.ghollami@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
pyperclip = "^1.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
